Move test cases to pinned Language version test file
